add rough module division
